// src/redux/types/actions.ts
export enum UserActionTypes {
  SET_AUTHENTICATED = 'SET_AUTHENTICATED',
  SET_UNAUTHENTICATED = 'SET_UNAUTHENTICATED',
  SET_USER = 'SET_USER',
  LOADING_USER = 'LOADING_USER',
}

export enum UIActionTypes {
  SET_ERRORS = 'SET_ERRORS',
  LOADING_UI = 'LOADING_UI',
  STOP_LOADING_UI = 'STOP_LOADING_UI',
  CLEAR_ERRORS = 'CLEAR_ERRORS',
  LOADING_DATA = 'LOADING_DATA',
}

export enum DataActionTypes {
  SET_POSTS = 'SET_POSTS',
  SET_POST = 'SET_POST',
  LIKE_POST = 'LIKE_POST',
  UNLIKE_POST = 'UNLIKE_POST',
  DELETE_POST = 'DELETE_POST',
  ADD_POST = 'ADD_POST',
  SUBMIT_COMMENT = 'SUBMIT_COMMENT',
}